+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [Add support for network filter option `top=`](https://github.com/gorhill/uBlock/commit/aebc108e36)
 - [Fix JS code extraction from web accessible resources](https://github.com/gorhill/uBlock/commit/a8bbd1a466)
 - [Add `freeze-element-property` scriptlet](https://github.com/gorhill/uBlock/commit/05f01f6be4)
 - [chromium][Categorize `.svg` resources as image type](https://github.com/gorhill/uBlock/commit/b862b73134)